Project import generated by Copybara.
GitOrigin-RevId: ae34f50469033a06f145c1313d2f0586947b01b5

+++ wallet-sdk/src/commonMain/kotlin/com/lightspark/sdk/wallet/graphql/RequestWithdrawal.kt
@@ -4,7 +4,7 @@ import com.lightspark.sdk.wallet.model.WithdrawalRequest
 
 const val RequestWithdrawalMutation = """
   mutation RequestWithdrawal(
-    ${'$'}amount_sats: Int!
+    ${'$'}amount_sats: Long!
     ${'$'}bitcoin_address: String!
   ) {
     request_withdrawal(input: {
